matrix request process
