+++ .gitignore
@@ -1,1 +1,3 @@
 obj
+*.sln
+.vs
