+++ Controls/NPImageMatrix.xaml
@@ -1,0 +1,78 @@
+﻿<UserControl x:Class="NPGui.Controls.NPImageMatrix"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:NPGui.Controls"
+             mc:Ignorable="d" 
+             d:DesignHeight="300" d:DesignWidth="300">
+    <Grid>
+        <Grid>
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition />
+                <ColumnDefinition Width="200"/>
+            </Grid.ColumnDefinitions>
+            <!-- IMAGE -->
+            <local:NPipeImageControl x:Name="FilteredImage" Grid.Column="0" />
+            <!-- PARAMETARS -->
+            <Grid Grid.Column="1">
+                <Grid.RowDefinitions>
+                    <RowDefinition Height="130"/>
+                    <RowDefinition Height="150" />
+                </Grid.RowDefinitions>
+                <!-- factor, bias, mat size -->
+                <Grid Grid.Row="0">
+                    <Grid.ColumnDefinitions>
+                        <ColumnDefinition Width="70"/>
+                        <ColumnDefinition />
+                    </Grid.ColumnDefinitions>
+                    <Grid.RowDefinitions>
+                        <RowDefinition Height="23" />
+                        <RowDefinition Height="23" />
+                        <RowDefinition Height="23" />
+                        <RowDefinition Height="23" />
+                        <RowDefinition Height="23" />
+                        <RowDefinition Height="23" />
+                        <RowDefinition Height="23" />
+                    </Grid.RowDefinitions>
+                    <!-- FACTOR -->
+                    <Label Content="factor:"                Grid.Row="0" Grid.Column="0" />
+                    <TextBox Name="factorTxt" Text="1.0"    Grid.Row="0" Grid.Column="1"/>
+                    <!-- BIAS -->
+                    <Label Content="bias:"                  Grid.Row="1" Grid.Column="0" />
+                    <TextBox Name="biasTxt" Text="0.0"      Grid.Row="1" Grid.Column="1"/>
+                    <!-- MATRIX WIDTH -->
+                    <Label Content="matWidth:"              Grid.Row="2" Grid.Column="0" />
+                    <TextBox Name="matWidthTxt" Text="3"    Grid.Row="2" Grid.Column="1"/>
+                    <!-- MATRIX HEIGHT -->
+                    <Label Content="matHeight:"             Grid.Row="3" Grid.Column="0" />
+                    <TextBox Name="matHeightTxt" Text="3" Grid.Row="3" Grid.Column="1"/>
+                    <!-- PROCESS -->
+                    <Button Name="processBtn" Content="Process" Click="processBtn_Click" Grid.Row="4" />
+                </Grid>
+                <!-- matrix 9x9 -->
+                <Grid Grid.Row="1">
+                    <Grid.ColumnDefinitions>
+                        <ColumnDefinition />
+                        <ColumnDefinition />
+                        <ColumnDefinition />
+                    </Grid.ColumnDefinitions>
+                    <Grid.RowDefinitions>
+                        <RowDefinition />
+                        <RowDefinition />
+                        <RowDefinition />
+                    </Grid.RowDefinitions>
+                    <TextBox Name="m00" Text="0.0"  Grid.Row="0" Grid.Column="0" />
+                    <TextBox Name="m01" Text="1.0"  Grid.Row="0" Grid.Column="1" />
+                    <TextBox Name="m02" Text="0.0"  Grid.Row="0" Grid.Column="2" />
+                    <TextBox Name="m10" Text="1.0"  Grid.Row="1" Grid.Column="0" />
+                    <TextBox Name="m11" Text="-4.0"  Grid.Row="1" Grid.Column="1" />
+                    <TextBox Name="m12" Text="1.0"  Grid.Row="1" Grid.Column="2" />
+                    <TextBox Name="m20" Text="0.0"  Grid.Row="2" Grid.Column="0" />
+                    <TextBox Name="m21" Text="1.0"  Grid.Row="2" Grid.Column="1" />
+                    <TextBox Name="m22" Text="0.0"  Grid.Row="2" Grid.Column="2" />
+                </Grid>
+            </Grid>
+        </Grid>
+    </Grid>
+</UserControl>
