+++ Controls/NPipeImageControl.xaml
@@ -1,0 +1,30 @@
+﻿<UserControl x:Class="NPGui.Controls.NPipeImageControl"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:NPGui.Controls"
+             mc:Ignorable="d" 
+             d:DesignHeight="300" d:DesignWidth="300">
+
+        <Grid Background="#FFE5E5E5" >
+            <Grid.RowDefinitions>
+                <RowDefinition Height="30"/>
+                <RowDefinition />
+            </Grid.RowDefinitions>
+            <ScrollViewer HorizontalScrollBarVisibility="Visible" Grid.Row="1"  >
+                <!-- IMAGE CONTROLL WITH NAMED PIPE-->
+            <Image x:Name="image" VerticalAlignment="Stretch" HorizontalAlignment="Stretch"/>
+        </ScrollViewer>
+            <Grid Grid.Row="0">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="70" />
+                    <ColumnDefinition Width="70" />
+                    <ColumnDefinition Width="70" />
+                </Grid.ColumnDefinitions>
+            </Grid>
+            <Button x:Name="processBtn" Content="Process" Click="processBtn_Click" HorizontalAlignment="Left" Margin="77,0,0,0" VerticalAlignment="Top" Width="75"/>
+            <Button x:Name="browseBtn" Content="Browse" Click="browseBtn_Click" HorizontalAlignment="Left" Margin="2,0,0,0" VerticalAlignment="Top" Width="75"/>
+        </Grid>
+        
+</UserControl>
